+++ test/test-osvvm.tcl
@@ -4,7 +4,7 @@
 # Note: tcllib is required for OSVVM
 
 # known good snapshot
-set OsvvmLibraries_tag "2022.10"
+set OsvvmLibraries_tag "2022.11"
 
 exec git clone --branch $OsvvmLibraries_tag --recurse-submodules -j8 https://github.com/OSVVM/OsvvmLibraries.git >@ stdout 2>@ stdout
 source OsvvmLibraries/Scripts/StartNVC.tcl