+++ test/test_parse.c
@@ -5080,6 +5080,33 @@ START_TEST(test_osvvm7)
 }
 END_TEST
 
+START_TEST(test_issue580)
+{
+   set_standard(STD_08);
+
+   input_from_file(TESTDIR "/parse/issue580.vhd");
+
+   const error_t expect[] = {
+      { 14, "declaration of signal S2 cannot have unconstrained type SUB_T" },
+      { -1, NULL }
+   };
+   expect_errors(expect);
+
+   tree_t e = parse();
+   fail_if(e == NULL);
+   fail_unless(tree_kind(e) == T_ENTITY);
+   lib_put(lib_work(), e);
+
+   tree_t a = parse();
+   fail_if(a == NULL);
+   fail_unless(tree_kind(a) == T_ARCH);
+
+   fail_unless(parse() == NULL);
+
+   check_expected_errors();
+}
+END_TEST
+
 Suite *get_parse_tests(void)
 {
    Suite *s = suite_create("parse");
@@ -5179,6 +5206,7 @@ Suite *get_parse_tests(void)
    tcase_add_test(tc_core, test_issue568);
    tcase_add_test(tc_core, test_issue569);
    tcase_add_test(tc_core, test_osvvm7);
+   tcase_add_test(tc_core, test_issue580);
    suite_add_tcase(s, tc_core);
 
    return s;